Update baseline class for 2017
