+++ Tools/Makefile
@@ -4,10 +4,11 @@ SDIR       = .
 
 CXX        = g++
 
-#CXXFLAGS  += -I. -std=c++0x -I$(CMSSW_BASE)/src/ -I$(LHAPDF_DATA_PATH)/../../include -I$(CMSSW_RELEASE_BASE)/../../../external/fastjet/3.1.0-odfocd/include -I$(CMSSW_RELEASE_BASE)/src
-CXXFLAGS  += -I. -std=c++0x -I$(CMSSW_BASE)/src/ -I$(LHAPDF_DATA_PATH)/../../include -I$(CMSSW_RELEASE_BASE)/src -I$(shell scram tool info fastjet | grep '^INCLUDE' | sed 's/INCLUDE=//')
+CXXFLAGS += -I. -std=c++0x -I$(CMSSW_BASE)/src/ -I$(LHAPDF_DATA_PATH)/../../include -I$(CMSSW_RELEASE_BASE)/src 
+CXXFLAGS += -I$(shell scram tool info fastjet | grep '^INCLUDE' | sed 's/INCLUDE=//')
+CXXFLAGS += -I$(shell scram tool info boost | grep '^INCLUDE' | sed 's/INCLUDE=//')
 ## Optimization flag
-CXXFLAGS += -g
+CXXFLAGS += -g -fPIC
 ## Enable the maximun warning
 #CXXFLAGS += -Wall -Wextra -Weffc++ -g
 
@@ -19,18 +20,29 @@ CXXDEPFLAGS = -MMD -MP
 LD         = g++
 LDFLAGS    =
 
+OPENCV_DIRECTORY=$(CMSSW_BASE)/src/opencv
 LIBS       = $(shell root-config --glibs)
+LIBS      += -Lobj -lSusyAnaTools
+LIBS      += -L$(CMSSW_BASE)/src/TopTagger/TopTagger/test -lTopTagger
+LIBS      += -L$(OPENCV_DIRECTORY)/lib/ -lopencv_ml -lopencv_core
 MT2LIB     = -L$(CMSSW_BASE)/lib/${SCRAM_ARCH}/ -lrecipeAUXOxbridgeMT2 -L$(CMSSW_RELEASE_BASE)/lib/${SCRAM_ARCH}/ -lPhysicsToolsHeppy
 LHAPDFLIB  = -L$(LHAPDF_DATA_PATH)/../../lib -lLHAPDF
-#OBJS       = $(patsubst %, $(ODIR)/%, $(OBJ))
+FILES = $(wildcard $(SDIR)/*.cc)
+OBJS  := $(FILES:$(SDIR)/%.cc=$(ODIR)/%.o)
 #PROGRAMS = tupleTest nEvts basicCheck makeCombPlots makeSignalHistograms signalScan makeSignalCards batchSignalPlots
 PROGRAMS = tupleTest nEvts makeCombPlots makeSignalHistograms signalScan basicCheck makeSignalCards makeUnblindPlots
 
-all: mkobj sampPyWrap $(PROGRAMS)
+#all: mkobj sampPyWrap  $(OBJS) $(PROGRAMS)
+all: mkobj $(OBJS) SusyAnaTools sampPyWrap   tupleTest 
 
 
 mkobj:
 	@mkdir -p obj
+
+
+SusyAnaTools : $(OBJS)
+	ar rcs lib$@.a $^
+	gcc -shared -o $(ODIR)/lib$@.so $^
 
 #code to compile shared library to link samples to python
 sampPyWrap: $(ODIR)/samplesModule.so
@@ -81,6 +93,6 @@ makeUnblindPlots: $(ODIR)/makeUnblindPlots.o $(ODIR)/searchBins.o
 	$(LD) $^ $(LIBS) $(MT2LIB) -o $@
 
 clean:
-	rm -f $(ODIR)/*.o $(ODIR)/*.d $(PROGRAMS) core 
+	rm -f $(ODIR)/*.so $(ODIR)/*.o $(ODIR)/*.d $(PROGRAMS) core 
 
 -include $(ODIR)/*.d
