+++ Tools/NTupleReader.h
@@ -87,6 +87,8 @@ public:
     bool getNextEvent();
     void disableUpdate();
     void printTupleMembers(FILE *f = stdout) const;
+    std::vector<std::string> GetTupleMembers() const;
+    std::vector<std::string> GetTupleSpecs(std::string VarName = "cntNJetsPt30Eta24") const;
 
     template<typename T> void registerFunction(T f)
     {
