+++ Tools/SATException.cc
